pyproject.toml
[tool.poetry]
name = "embodiedmind"
version = "0.1.0"
description = "LangChain-powered embodied intelligence knowledge agent"
authors = []
readme = "README.md"
packages = [{include = "embodiedmind", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = ">=0.3.0"
langchain-openai = ">=0.2.0"
langchain-community = ">=0.3.0"
langchain-chroma = ">=0.1.0"
chromadb = ">=0.5.0"
openai = ">=1.0.0"
gradio = ">=4.0.0"
fastapi = ">=0.110.0"
uvicorn = ">=0.29.0"
playwright = ">=1.44.0"
beautifulsoup4 = ">=4.12.0"
apscheduler = ">=3.10.0"
pydantic-settings = ">=2.0.0"
tavily-python = ">=0.3.0"
arxiv = ">=2.1.0"
pypdf = ">=4.0.0"
httpx = ">=0.27.0"
tenacity = ">=8.3.0"
rich = ">=13.0.0"
PyGithub = ">=2.3.0"
gitpython = ">=3.1.0"

[tool.poetry.dev-dependencies]
pytest = ">=8.0"
pytest-asyncio = ">=0.23"
ruff = ">=0.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
